@@ -1,35 +1,38 @@
 {
-  "name": "react-hooks-demo",
-  "version": "1.0.0",
-  "description": "",
-  "keywords": [],
-  "homepage": "https://ms0372631.github.io/Todo-List",
-  "main": "src/index.js",
+  "name": "movie-page",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "@babel/runtime": "7.13.8",
-    "gh-pages": "^4.0.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-redux": "^8.0.2",
-    "react-scripts": "4.0.0",
-    "redux": "^4.2.0"
-  },
-  "devDependencies": {
-    "@redux-devtools/core": "^3.13.1",
-    "typescript": "4.1.3"
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^2.1.3",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
